+++ app/src/main/java/com/sahariar/star/crickscorer/database/TeamDB.java
@@ -3,6 +3,7 @@ package com.sahariar.star.crickscorer.database;
 import android.content.ContentValues;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
+import android.util.Log;
 
 import com.sahariar.star.crickscorer.Model.PlayerModel;
 import com.sahariar.star.crickscorer.Model.Team;
@@ -36,12 +37,21 @@ public class TeamDB {
     }
     public boolean delete(SQLiteDatabase db,int id)
     {
+       MatchDB m=new MatchDB();
 
         try {
-            db.delete(tableName, "id=" + id, null);
-            PlayerToTeam ptm=new PlayerToTeam();
-            ptm.delte(db,id);
-            return ptm.delte(db,id);
+             if(m.deleteByTeamId(db,id)) {
+
+
+                 db.delete(tableName, "id=" + id, null);
+                 PlayerToTeam ptm = new PlayerToTeam();
+                 ptm.delte(db, id);
+                 return ptm.delte(db, id);
+             }
+             else
+             {
+                 return false;
+             }
         }
         catch(Exception e)
         {
@@ -50,7 +60,8 @@ public class TeamDB {
     }
     public String getTeamName(SQLiteDatabase db,long id)
     {
-        String sql="Select * from "+tableName+" where id="+id;
+        String sql="Select * from "+tableName+" WHERE id = "+id;
+
         Cursor res=db.rawQuery(sql,null);
         res.moveToFirst();
 
@@ -59,6 +70,7 @@ public class TeamDB {
         while(!res.isAfterLast())
         {
 
+
             team=new Team(Integer.parseInt(res.getString(0)),res.getString(1));
 
             res.moveToNext();
@@ -66,10 +78,31 @@ public class TeamDB {
 
         return team.getName();
     }
+    public Team getOne(SQLiteDatabase db,long id)
+    {
+        String sql="Select * from "+tableName+" WHERE id = "+id;
+
+        Cursor res=db.rawQuery(sql,null);
+        res.moveToFirst();
+
+        Team team=new Team(0,null);
+
+        while(!res.isAfterLast())
+        {
+
+
+            team=new Team(Integer.parseInt(res.getString(0)),res.getString(1));
+
+            res.moveToNext();
+        }
+
+        return team;
+    }
+
 
     //remove team
     //show all team
-    public List<Team> getAllplayers(SQLiteDatabase db)
+    public List<Team> getAllTeams(SQLiteDatabase db)
     {
         String sql="Select * from "+tableName;
         Cursor res=db.rawQuery(sql,null);
